+++ cmd/zephyr/titlebar_windows.go
@@ -21,10 +21,6 @@ func pointerOutsideWindow() bool { return false }
 
 func startWindowDrag() {}
 
-func globalMousePosition() (x, y float64) { return 0, 0 }
-
-func windowFrame() (x, y, w, h float64) { return 0, 0, 0, 0 }
-
 const trafficLightPaddingDp = 0
 
 func updateWindowBackground(c color.NRGBA) {}